+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Added ec2 vpc endpoints support
  - Added ec2 elastic IP support
  - Added elbv2 target group support
+ - Added elbv2 load balancer support
 
 
 # 0.16.1